requirements.txt
# FastAPI and Server
fastapi[standard]>=0.115.8

# Authentication
fastapi-azure-auth>=5.0.0
python-jose[cryptography]>=3.3.0

# Database
motor>=3.7.0
pymongo>=4.11.1

# Cloud Services
google-cloud-tasks>=2.19.0
google-auth>=2.38.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.6
pydantic-settings>=2.6.1
email-validator>=2.2.0

# Monitoring and Logging
prometheus-fastapi-instrumentator>=7.0.2
structlog>=25.1.0
python-json-logger>=3.2.1

# Development Tools
black>=25.1.0
isort>=6.0.0
flake8>=7.1.2
pytest>=8.3.4
pytest-asyncio>=0.25.3
httpx>=0.28.1

# Production Dependencies
tenacity>=9.0.0
aiohttp>=3.11.12